requirements.txt
scikit-learn
faiss-cpu
sentence-transformers
gensim
streamlit
